+++ source/include/rocprofiler-sdk/fwd.h
@@ -572,10 +572,9 @@ typedef union rocprofiler_address_t
  * @brief Stores UUID for devices.
  *
  */
-typedef union rocprofiler_uuid_t
+typedef struct rocprofiler_uuid_t
 {
-    uint64_t value;  ///< numerical value
-    void*    bytes;  ///< uuid in hexadecimal
+    uint8_t bytes[16];  // numerical value
 } rocprofiler_uuid_t;
 
 //--------------------------------------------------------------------------------------//