+++ CHANGELOG.md
@@ -198,6 +198,7 @@ Full documentation for ROCprofiler-SDK is available at [rocm.docs.amd.com/projec
 
 - SDK no longer creates a background thread when every tool returns a nullptr from `rocprofiler_configure`.
 - Updated disassembly.hpp's vaddr-to-file-offset mapping to use the dedicated comgr API.
+- rocprofiler_uuid_t ABI is changed to hold 128 bit value.
 - rocprofv3 shorthand argument for `--collection-period` is now `-P` (upper-case) as `-p` (lower-case) is reserved for later use
 - default output format for rocprofv3 is now `rocpd` (SQLite3 database)
 - rocprofv3 avail tool renamed from rocprofv3_avail to rocprofv3-avail tool